MISUV-8334 | Moved ChargeItem to own file
